requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
motor>=3.3
pymongo>=4.6
pydantic>=2.6
python-dotenv>=1.0